Unify getting aws credentials
